Create observation package with ObservationInternal
