+++ internal/server/handler_core.go
@@ -22,8 +22,6 @@ import (
 	v1e "github.com/datacommonsorg/mixer/internal/server/v1/event"
 	v2 "github.com/datacommonsorg/mixer/internal/server/v2"
 	v2e "github.com/datacommonsorg/mixer/internal/server/v2/event"
-	v2facet "github.com/datacommonsorg/mixer/internal/server/v2/facet"
-	v2observation "github.com/datacommonsorg/mixer/internal/server/v2/observation"
 	v2p "github.com/datacommonsorg/mixer/internal/server/v2/properties"
 	v2pv "github.com/datacommonsorg/mixer/internal/server/v2/propertyvalues"
 	"github.com/datacommonsorg/mixer/internal/server/v2/resolve"
@@ -252,131 +250,4 @@ func (s *Server) V2EventCore(
 	}
 
 	return nil, nil
-}
-
-// V2ObservationCore fetches observation from Cloud Bigtable.
-func (s *Server) V2ObservationCore(
-	ctx context.Context, in *pbv2.ObservationRequest,
-) (*pbv2.ObservationResponse, error) {
-	// (TODO): The routing logic here is very rough. This needs more work.
-	var queryDate, queryValue, queryVariable, queryEntity, queryFacet bool
-	for _, item := range in.GetSelect() {
-		if item == "date" {
-			queryDate = true
-		} else if item == "value" {
-			queryValue = true
-		} else if item == "variable" {
-			queryVariable = true
-		} else if item == "entity" {
-			queryEntity = true
-		} else if item == "facet" {
-			queryFacet = true
-		}
-	}
-	if !queryVariable || !queryEntity {
-		return nil, status.Error(
-			codes.InvalidArgument, "Must select 'variable' and 'entity'")
-	}
-
-	variable := in.GetVariable()
-	entity := in.GetEntity()
-
-	// Observation date and value query.
-	if queryDate && queryValue {
-		// Series.
-		if len(variable.GetDcids()) > 0 && len(entity.GetDcids()) > 0 {
-			return v2observation.FetchDirect(
-				ctx,
-				s.store,
-				s.cachedata.Load().SQLProvenances(),
-				variable.GetDcids(),
-				entity.GetDcids(),
-				in.GetDate(),
-				in.GetFilter(),
-			)
-		}
-
-		// Collection.
-		if len(variable.GetDcids()) > 0 && entity.GetExpression() != "" {
-			// Example of expression
-			// "geoId/06<-containedInPlace+{typeOf: City}"
-			expr := entity.GetExpression()
-			containedInPlace, err := v2.ParseContainedInPlace(expr)
-			if err != nil {
-				return nil, err
-			}
-			return v2observation.FetchContainedIn(
-				ctx,
-				s.store,
-				s.metadata,
-				s.cachedata.Load().SQLProvenances(),
-				s.httpClient,
-				s.metadata.RemoteMixerDomain,
-				variable.GetDcids(),
-				containedInPlace.Ancestor,
-				containedInPlace.ChildPlaceType,
-				in.GetDate(),
-				in.GetFilter(),
-			)
-		}
-
-		// Derived series.
-		if variable.GetFormula() != "" && len(entity.GetDcids()) > 0 {
-			return v2observation.DerivedSeries(
-				ctx,
-				s.store,
-				variable.GetFormula(),
-				entity.GetDcids(),
-			)
-		}
-	}
-
-	// Get facet information for <variable, entity> pair.
-	if !queryDate && !queryValue && queryFacet {
-		// Series
-		if len(variable.GetDcids()) > 0 && len(entity.GetDcids()) > 0 {
-			return v2facet.SeriesFacet(
-				ctx,
-				s.store,
-				s.cachedata.Load(),
-				variable.GetDcids(),
-				entity.GetDcids(),
-			)
-		}
-		// Collection
-		if len(variable.GetDcids()) > 0 && entity.GetExpression() != "" {
-			expr := entity.GetExpression()
-			containedInPlace, err := v2.ParseContainedInPlace(expr)
-			if err != nil {
-				return nil, err
-			}
-			return v2facet.ContainedInFacet(
-				ctx,
-				s.store,
-				s.cachedata.Load(),
-				s.metadata,
-				s.httpClient,
-				s.metadata.RemoteMixerDomain,
-				variable.GetDcids(),
-				containedInPlace.Ancestor,
-				containedInPlace.ChildPlaceType,
-				in.GetDate(),
-			)
-		}
-	}
-
-	// Get existence of <variable, entity> pair.
-	if !queryDate && !queryValue {
-		if len(entity.GetDcids()) > 0 {
-			if len(variable.GetDcids()) > 0 {
-				// Have both entity.dcids and variable.dcids. Check existence cache.
-				return v2observation.Existence(
-					ctx, s.store, s.cachedata.Load(), variable.GetDcids(), entity.GetDcids())
-			}
-			// TODO: Support appending entities from entity.expression
-			// Only have entity.dcids, fetch variables for each entity.
-			return v2observation.Variable(ctx, s.store, entity.GetDcids())
-		}
-	}
-	return &pbv2.ObservationResponse{}, nil
 }
